lib/i915: Use explicit integer casts to satisfy clang
../lib/i915/gem_engine_topology.c:123:15: warning: result of comparison of constant -2 with expression of type 'uint16_t' (aka 'unsigned short') is always false [-Wtautological-constant-out-of-range-compare]
            instance == I915_ENGINE_CLASS_INVALID_VIRTUAL) {
            ~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../lib/i915/gem_engine_topology.c:122:12: warning: result of comparison of constant 'I915_ENGINE_CLASS_INVALID' (-1) with expression of type 'uint16_t' (aka 'unsigned short') is always false [-Wtautological-constant-out-of-range-compare]
        if (class == I915_ENGINE_CLASS_INVALID &&
            ~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~

Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>

+++ lib/i915/gem_engine_topology.c
@@ -119,8 +119,8 @@ static void init_engine(struct intel_execution_engine2 *e2,
 	e2->instance = instance;
 
 	/* engine is a virtual engine */
-	if (class == I915_ENGINE_CLASS_INVALID &&
-	    instance == I915_ENGINE_CLASS_INVALID_VIRTUAL) {
+	if (class == (uint16_t)I915_ENGINE_CLASS_INVALID &&
+	    instance == (uint16_t)I915_ENGINE_CLASS_INVALID_VIRTUAL) {
 		strcpy(e2->name, "virtual");
 		e2->is_virtual = true;
 		return;
